# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Transferências e Usuários",
        "version": "1.0.0",
        "description": "API para registro, login, consulta de usuários e transferências."
    },
    "paths": {
        "/users/register": {
            "post": {
                "summary": "Registrar novo usuário",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "favorecidos": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "required": [
                                    "username",
                                    "password"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário registrado"
                    },
                    "400": {
                        "description": "Usuário já existe"
                    }
                }
            }
        },
        "/users/login": {
            "post": {
                "summary": "Login de usuário",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "username",
                                    "password"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado"
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        },
        "/users": {
            "get": {
                "summary": "Consultar usuários",
                "responses": {
                    "200": {
                        "description": "Lista de usuários"
                    }
                }
            }
        },
        "/transfer": {
            "post": {
                "summary": "Realizar transferência",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "from": {
                                        "type": "string"
                                    },
                                    "to": {
                                        "type": "string"
                                    },
                                    "amount": {
                                        "type": "number"
                                    }
                                },
                                "required": [
                                    "from",
                                    "to",
                                    "amount"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transferência realizada"
                    },
                    "400": {
                        "description": "Erro na transferência"
                    }
                }
            },
            "get": {
                "summary": "Consultar transferências",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de transferências"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}